feat: refactor mutant test cases
