+++ packages/core/src/__tests__/Spore.test.ts
@@ -223,86 +223,4 @@ describe('Spore', () => {
       ).rejects.toThrow();
     }, 0);
   });
-
-  describe('Spore with Mutant', () => {
-    it('Create a Spore with Mutant', async () => {
-      const immortalMutantId = '0x79713beaf43310d4d9c838811553399f3e7c114353d4788de3ed2a165e288c11';
-
-      const { txSkeleton, outputIndex, reference, mutantReference } = await createSpore({
-        data: {
-          contentType: 'text/plain',
-          content: bytifyRawString('content'),
-          contentTypeParameters: {
-            mutant: [immortalMutantId],
-          },
-        },
-        fromInfos: [CHARLIE.address],
-        toLock: CHARLIE.lock,
-        config,
-      });
-
-      const spore = getSporeOutput(txSkeleton, outputIndex, config);
-      expect(spore.data.contentType).toEqual(`text/plain;mutant[]=${immortalMutantId.slice(2)}`);
-
-      expect(reference).toBeDefined();
-      expect(reference.referenceTarget).toEqual('none');
-
-      expect(mutantReference).toBeDefined();
-      const mutantCell = await getMutantById(immortalMutantId, config);
-      const mutantArgs = unpackToRawMutantArgs(mutantCell.cellOutput.type!.args);
-      if (mutantArgs.minPayment !== void 0) {
-        expect(mutantReference!.referenceType).toEqual('payment');
-        expect(mutantReference!.payment).toBeDefined();
-        expect(mutantReference!.payment!.outputIndices.length).toEqual(1);
-        const paymentCell = txSkeleton.get('outputs').get(mutantReference!.payment!.outputIndices[0]);
-        expect(BI.from(paymentCell!.cellOutput.capacity).gte(BI.from(mutantArgs.minPayment))).toEqual(true);
-        expect(paymentCell!.cellOutput.lock).toEqual(mutantCell.cellOutput.lock);
-      } else {
-        expect(mutantReference!.referenceType).toEqual('none');
-      }
-
-      const hash = await signAndSendTransaction({
-        account: CHARLIE,
-        txSkeleton,
-        config,
-        rpc,
-        send: true,
-      });
-      if (hash) {
-        SPORE_OUTPOINT_RECORDS.push({
-          outPoint: {
-            txHash: hash,
-            index: BI.from(outputIndex).toHexString(),
-          },
-          account: CHARLIE,
-        });
-      }
-    }, 0);
-    it('Create a Spore with Mutant required Cluster', async () => {
-      const { txSkeleton, outputIndex, reference, mutantReference } = await createSpore({
-        data: {
-          contentType: 'text/plain',
-          content: bytifyRawString('content'),
-          clusterId: '0x',
-          contentTypeParameters: {
-            mutant: ['0x'],
-          },
-        },
-        fromInfos: [CHARLIE.address],
-        toLock: CHARLIE.lock,
-        config,
-      });
-
-      console.log('Spore Reference:', reference);
-      console.log('Spore MutantReference:', mutantReference);
-
-      await signAndSendTransaction({
-        account: CHARLIE,
-        txSkeleton,
-        config,
-        rpc,
-        send: false,
-      });
-    }, 0);
-  });
 });
