+++ packages/core/src/__tests__/resources/immortalMutant.lua
@@ -1,3 +1,4 @@
+-- Melt a Spore denied
 if spore_ext_mode == 3 then
-  ckb.exit_script("86")
+  ckb.exit_script("99")
 end
